move put handler back to web_scraping.js
